Endevender as a virtual dimmer
Reports of switch and level cluster are used to update a virtual dimmer device. This device can again be synchronised to another dimmer to achieve desired functionality.

Please be warned that the endevender only seems to report one change per second, so for the time being this is very slow.